#  optimize Photoneo YCoCg-R 420 color conversion