--- a/include/camera_aravis/conversion_utils.h
+++ b/include/camera_aravis/conversion_utils.h
@@ -163,7 +163,7 @@ const std::map<std::string, ConversionFunction> CONVERSIONS_DICTIONARY =
  //// data adapters
  { "FloatToUint", std::bind(&float_to_uint, std::placeholders::_1, std::placeholders::_2, 1.0f, sensor_msgs::image_encodings::TYPE_16UC1) },
  //// crazy internal pixel formats fixing various device quirks
- { "PhotoneoYCoCg420", std::bind(&photoneoYCoCgR420, std::placeholders::_1, std::placeholders::_2, sensor_msgs::image_encodings::RGB8) }
+ { "PhotoneoYCoCg420", std::bind(&photoneoYCoCgR420, std::placeholders::_1, std::placeholders::_2, sensor_msgs::image_encodings::BGRA8) }
 };
 
 } // end namespace camera_aravis