In #13 we investigated that data sent from device is in `YCoCg-R`  pixel format 
- not really "custom `YCoCg`" as documented

Optimize the implementation by:
- following standard VESA DSC formulation of algorithm
- and process in row pointer logic

This yields:
- power-safe (standard) mode of my laptop max 30 ms to max 11 ms (high variability as clocks speed change)
- in performance mode from ~7 ms to ~3 ms

Finally, instead of converting to RGB
- convert to BGRA8
- which maps directly to VAAPI BGR0
- which prevents another color conversion when hardware encoding
- and saves a few % of CPU usage of my laptop
  - mostly important for embedded low power devices

----------------------------------------------

Further optimizations are possible and relatively simple with SIMD
- we are already processing groups of 2x2 pixels in SIMD friendly manner